Move location setup to samples .cpp
