+++ Source/CesiumRuntime/Private/Tests/CesiumSceneGeneration.cpp
@@ -11,7 +11,6 @@
 
 #include "Cesium3DTileset.h"
 #include "CesiumGeoreference.h"
-#include "CesiumIonRasterOverlay.h"
 #include "CesiumSunSky.h"
 #include "GlobeAwareDefaultPawn.h"
 
@@ -158,65 +157,6 @@ void createCommonWorldObjects(SceneGenerationContext& context) {
   AWorldSettings* pWorldSettings = context.world->GetWorldSettings();
   if (pWorldSettings)
     pWorldSettings->bEnableWorldBoundsChecks = false;
-}
-
-void setupForDenver(SceneGenerationContext& context) {
-
-  context.setCommonProperties(
-      FVector(-104.988892, 39.743462, 1798.679443),
-      FVector(0, 0, 0),
-      FRotator(-5.2, -149.4, 0),
-      90.0f);
-
-  // Add Cesium World Terrain
-  ACesium3DTileset* worldTerrainTileset =
-      context.world->SpawnActor<ACesium3DTileset>();
-  worldTerrainTileset->SetTilesetSource(ETilesetSource::FromCesiumIon);
-  worldTerrainTileset->SetIonAssetID(1);
-  worldTerrainTileset->SetIonAccessToken(SceneGenerationContext::testIonToken);
-  worldTerrainTileset->SetActorLabel(TEXT("Cesium World Terrain"));
-
-  // Bing Maps Aerial overlay
-  UCesiumIonRasterOverlay* pOverlay = NewObject<UCesiumIonRasterOverlay>(
-      worldTerrainTileset,
-      FName("Bing Maps Aerial"),
-      RF_Transactional);
-  pOverlay->MaterialLayerKey = TEXT("Overlay0");
-  pOverlay->IonAssetID = 2;
-  pOverlay->SetActive(true);
-  pOverlay->OnComponentCreated();
-  worldTerrainTileset->AddInstanceComponent(pOverlay);
-
-  // Aerometrex Denver
-  ACesium3DTileset* aerometrexTileset =
-      context.world->SpawnActor<ACesium3DTileset>();
-  aerometrexTileset->SetTilesetSource(ETilesetSource::FromCesiumIon);
-  aerometrexTileset->SetIonAssetID(354307);
-  aerometrexTileset->SetIonAccessToken(SceneGenerationContext::testIonToken);
-  aerometrexTileset->SetMaximumScreenSpaceError(2.0);
-  aerometrexTileset->SetActorLabel(TEXT("Aerometrex Denver"));
-
-  context.tilesets.push_back(worldTerrainTileset);
-  context.tilesets.push_back(aerometrexTileset);
-}
-
-void setupForMontrealPointCloud(SceneGenerationContext& context) {
-
-  context.setCommonProperties(
-      FVector(-73.616526, 45.57335, 95.048859),
-      FVector(0, 0, 0),
-      FRotator(-90.0, 0.0, 0.0),
-      90.0f);
-
-  ACesium3DTileset* montrealTileset =
-      context.world->SpawnActor<ACesium3DTileset>();
-  montrealTileset->SetTilesetSource(ETilesetSource::FromCesiumIon);
-  montrealTileset->SetIonAssetID(28945);
-  montrealTileset->SetIonAccessToken(SceneGenerationContext::testIonToken);
-  montrealTileset->SetMaximumScreenSpaceError(16.0);
-  montrealTileset->SetActorLabel(TEXT("Montreal Point Cloud"));
-
-  context.tilesets.push_back(montrealTileset);
 }
 
 } // namespace Cesium
